refactor: tidy, move and rename allinone
