+++ src/service/account-service.ts
@@ -14,7 +14,7 @@ import {
   kaspaToSompi,
   ITransactionOutput,
 } from "kaspa-wasm";
-import { KaspaClient } from "../utils/all-in-one";
+import { KaspaClient } from "../utils/kaspa-client";
 import { encrypt_message } from "cipher";
 import { DecryptionCache } from "../utils/decryption-cache";
 import { CipherHelper } from "../utils/cipher-helper";
